los: move to trx; go with TR2+ impl
