+++ src/tr1/meson.build
@@ -130,7 +130,6 @@ sources = [
   'game/item_actions/stairs2slope.c',
   'game/item_actions/turn_180.c',
   'game/lara/common.c',
-  'game/los.c',
   'game/objects/common.c',
   'game/objects/creatures/cutscene_player.c',
   'game/objects/general/earthquake.c',
