--- src/tr2/game/los.h
@@ -1,3 +1,0 @@
-#pragma once
-
-#include <libtrx/game/los.h>
