+++ src/tr2/meson.build
@@ -111,7 +111,6 @@ sources = [
   'game/lara/flare.c',
   'game/lara/misc.c',
   'game/level.c',
-  'game/los.c',
   'game/objects/common.c',
   'game/objects/creatures/skidoo_driver.c',
   'game/objects/general/combat_end.c',
